Move all tests into folder contents
