+++ tests/favorites/create-favs.spec.js
@@ -1,10 +1,10 @@
 require('dotenv').config();
 var request = require("supertest");
-var app = require('../app');
+var app = require('../../app');
 const nock = require('nock');
 
 const environment = process.env.NODE_ENV || 'test';
-const configuration = require('../knexfile')[environment];
+const configuration = require('../../knexfile')[environment];
 const database = require('knex')(configuration);
 
 describe('Test creating favorites', () => {
