+++ tests/favorites/del-favorites-by-id.spec.js
@@ -1,8 +1,8 @@
 var request = require("supertest");
-var app = require('../app');
+var app = require('../../app');
 
 const environment = process.env.NODE_ENV || 'test';
-const configuration = require('../knexfile')[environment];
+const configuration = require('../../knexfile')[environment];
 const database = require('knex')(configuration);
 
 describe('Test the favorites endpoints', () => {
